Improved CommitBuffer()
CommitBuffer applies a mutex within the Buffer however , if multiple context exist that write to the same file, there was no guarantee that the writes at the same instant would not mix. Added another mutex @ a target (in this case file) level to ensure un-related context to the same file also do so sequentially.